+++ src/main.tsx
@@ -5,15 +5,18 @@ import HomePage from './components/HomePage'
 import UploadScan from './components/UploadScan'
 import AnalysisReport from './components/AnalysisReport'
 import './index.css'
+import { ThemeProvider } from './theme/ThemeContext';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <Router>
-      <Routes>
-        <Route path="/" element={<HomePage />} />
-        <Route path="/upload" element={<UploadScan />} />
-        <Route path="/report" element={<AnalysisReport />} />
-      </Routes>
-    </Router>
+    <ThemeProvider>
+      <Router>
+        <Routes>
+          <Route path="/" element={<HomePage />} />
+          <Route path="/upload" element={<UploadScan />} />
+          <Route path="/report" element={<AnalysisReport />} />
+        </Routes>
+      </Router>
+    </ThemeProvider>
   </React.StrictMode>,
 )
